+++ flutter/lib/src/integrations/connectivity/connectivity_provider.dart
@@ -1,5 +1,6 @@
 import 'noop_connectivity_provider.dart'
-    if (dart.library.html) 'web_connectivity_provider.dart';
+    if (dart.library.html) 'html_connectivity_provider.dart'
+    if (dart.library.js_interop) 'web_connectivity_provider.dart';
 
 abstract class ConnectivityProvider {
   factory ConnectivityProvider() => connectivityProvider();