+++ flutter/lib/src/integrations/connectivity/web_connectivity_provider.dart
@@ -1,5 +1,8 @@
 import 'dart:async';
-import 'dart:html' as html;
+
+// We would lose compatibility with old dart versions by adding web to pubspec.
+// ignore: depend_on_referenced_packages
+import 'package:web/web.dart' as web;
 
 import 'connectivity_provider.dart';
 
@@ -8,15 +11,19 @@ ConnectivityProvider connectivityProvider() {
 }
 
 class WebConnectivityProvider implements ConnectivityProvider {
-  StreamSubscription<html.Event>? _onOnlineSub;
-  StreamSubscription<html.Event>? _onOfflineSub;
+  StreamSubscription<web.Event>? _onOnlineSub;
+  StreamSubscription<web.Event>? _onOfflineSub;
 
   @override
   void listen(void Function(String connectivity) onChange) {
-    _onOnlineSub = html.window.onOnline.listen((_) {
+    _onOnlineSub = web.EventStreamProviders.onlineEvent
+        .forElement(web.document.body!)
+        .listen((_) {
       onChange('wifi');
     });
-    _onOfflineSub = html.window.onOffline.listen((_) {
+    _onOfflineSub = web.EventStreamProviders.offlineEvent
+        .forElement(web.document.body!)
+        .listen((_) {
       onChange('none');
     });
   }