+++ flutter/test/integrations/load_contexts_integration_test.dart
@@ -1,39 +1,23 @@
 @TestOn('vm')
+library flutter_test;
 
-import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/integrations/load_contexts_integration.dart';
 
-import '../mocks.dart';
-import '../mocks.mocks.dart';
+import 'fixture.dart';
 
 void main() {
   group(LoadContextsIntegration, () {
-    const _channel = MethodChannel('sentry_flutter');
+    late IntegrationTestFixture<LoadContextsIntegration> fixture;
 
-    TestWidgetsFlutterBinding.ensureInitialized();
-
-    late Fixture fixture;
-
-    setUp(() {
-      fixture = Fixture();
-    });
-
-    tearDown(() {
-      // ignore: deprecated_member_use
-      _channel.setMockMethodCallHandler(null);
+    setUp(() async {
+      fixture = IntegrationTestFixture(LoadContextsIntegration.new);
+      await fixture.registerIntegration();
     });
 
     test('loadContextsIntegration adds integration', () {
-      // ignore: deprecated_member_use
-      _channel.setMockMethodCallHandler((MethodCall methodCall) async {});
-
-      final integration = LoadContextsIntegration(_channel);
-
-      integration(fixture.hub, fixture.options);
-
       expect(
           fixture.options.sdk.integrations.contains('loadContextsIntegration'),
           true);
@@ -45,19 +29,10 @@ void main() {
       final eventBreadcrumb = Breadcrumb(message: 'event');
       var event = SentryEvent(breadcrumbs: [eventBreadcrumb]);
 
-      final nativeBreadcrumb = Breadcrumb(message: 'native');
-      Map<String, dynamic> loadContexts = {
-        'breadcrumbs': [nativeBreadcrumb.toJson()]
-      };
+      when(fixture.binding.loadContexts()).thenAnswer((_) async => {
+            'breadcrumbs': [Breadcrumb(message: 'native').toJson()]
+          });
 
-      final future = Future.value(loadContexts);
-      when(fixture.methodChannel.invokeMethod<dynamic>('loadContexts'))
-          .thenAnswer((_) => future);
-      // ignore: deprecated_member_use
-      _channel.setMockMethodCallHandler((MethodCall methodCall) async {});
-
-      final integration = LoadContextsIntegration(fixture.methodChannel);
-      integration.call(fixture.hub, fixture.options);
       event =
           (await fixture.options.eventProcessors.first.apply(event, Hint()))!;
 
@@ -71,19 +46,10 @@ void main() {
       final eventBreadcrumb = Breadcrumb(message: 'event');
       var event = SentryEvent(breadcrumbs: [eventBreadcrumb]);
 
-      final nativeBreadcrumb = Breadcrumb(message: 'native');
-      Map<String, dynamic> loadContexts = {
-        'breadcrumbs': [nativeBreadcrumb.toJson()]
-      };
+      when(fixture.binding.loadContexts()).thenAnswer((_) async => {
+            'breadcrumbs': [Breadcrumb(message: 'native').toJson()]
+          });
 
-      final future = Future.value(loadContexts);
-      when(fixture.methodChannel.invokeMethod<dynamic>('loadContexts'))
-          .thenAnswer((_) => future);
-      // ignore: deprecated_member_use
-      _channel.setMockMethodCallHandler((MethodCall methodCall) async {});
-
-      final integration = LoadContextsIntegration(fixture.methodChannel);
-      integration.call(fixture.hub, fixture.options);
       event =
           (await fixture.options.eventProcessors.first.apply(event, Hint()))!;
 
@@ -104,35 +70,49 @@ void main() {
       final eventBreadcrumb = Breadcrumb(message: 'event');
       var event = SentryEvent(breadcrumbs: [eventBreadcrumb]);
 
-      final nativeMutatedBreadcrumb = Breadcrumb(message: 'native-mutated');
-      final nativeDeletedBreadcrumb = Breadcrumb(message: 'native-deleted');
-      Map<String, dynamic> loadContexts = {
-        'breadcrumbs': [
-          nativeMutatedBreadcrumb.toJson(),
-          nativeDeletedBreadcrumb.toJson(),
-        ]
-      };
+      when(fixture.binding.loadContexts()).thenAnswer((_) async => {
+            'breadcrumbs': [
+              Breadcrumb(message: 'native-mutated').toJson(),
+              Breadcrumb(message: 'native-deleted').toJson(),
+            ]
+          });
 
-      final future = Future.value(loadContexts);
-      when(fixture.methodChannel.invokeMethod<dynamic>('loadContexts'))
-          .thenAnswer((_) => future);
-      // ignore: deprecated_member_use
-      _channel.setMockMethodCallHandler((MethodCall methodCall) async {});
-
-      final integration = LoadContextsIntegration(fixture.methodChannel);
-      integration.call(fixture.hub, fixture.options);
       event =
           (await fixture.options.eventProcessors.first.apply(event, Hint()))!;
 
       expect(event.breadcrumbs!.length, 1);
       expect(event.breadcrumbs!.first.message, 'native-mutated-applied');
     });
+
+    test('apply default IP to user during captureEvent after loading context',
+        () async {
+      fixture.options.enableScopeSync = true;
+
+      const expectedIp = '{{auto}}';
+      String? actualIp;
+
+      const expectedId = '1';
+      String? actualId;
+
+      fixture.options.beforeSend = (event, hint) {
+        actualIp = event.user?.ipAddress;
+        actualId = event.user?.id;
+        return event;
+      };
+
+      final options = fixture.options;
+
+      final user = SentryUser(id: expectedId);
+      when(fixture.binding.loadContexts())
+          .thenAnswer((_) async => {'user': user.toJson()});
+
+      final client = SentryClient(options);
+      final event = SentryEvent();
+
+      await client.captureEvent(event);
+
+      expect(expectedIp, actualIp);
+      expect(expectedId, actualId);
+    });
   });
-}
-
-class Fixture {
-  final hub = MockHub();
-  final options = SentryFlutterOptions(dsn: fakeDsn);
-
-  final methodChannel = MockMethodChannel();
 }