+++ flutter/test/integrations/load_contexts_integrations_test.dart
@@ -1,7 +1,8 @@
 @TestOn('vm')
+library flutter_test;
 
-import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/mockito.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/integrations/load_contexts_integration.dart';
 
@@ -83,7 +84,7 @@ void main() {
 
     final event = await fixture.options.eventProcessors.first.apply(e, Hint());
 
-    expect(fixture.called, true);
+    verify(fixture.binding.loadContexts()).called(1);
     expect(event?.contexts.device?.name, 'Device1');
     expect(event?.contexts.app?.name, 'test-app');
     expect(event?.contexts.app?.inForeground, true);
@@ -123,7 +124,7 @@ void main() {
 
     final event = await fixture.options.eventProcessors.first.apply(e, Hint());
 
-    expect(fixture.called, true);
+    verify(fixture.binding.loadContexts()).called(1);
     expect(event?.contexts.device?.name, 'eDevice');
     expect(event?.contexts.app?.name, 'eApp');
     expect(event?.contexts.app?.inForeground, true);
@@ -235,10 +236,7 @@ void main() {
   );
 
   test('should not throw on loadContextsIntegration exception', () async {
-    // ignore: deprecated_member_use
-    fixture.channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      throw Exception();
-    });
+    when(fixture.binding.loadContexts()).thenThrow(Exception());
     final integration = fixture.getSut();
     integration(fixture.hub, fixture.options);
 
@@ -425,12 +423,9 @@ void main() {
 }
 
 class Fixture {
-  final channel = MethodChannel('sentry_flutter');
-
   final hub = MockHub();
   final options = SentryFlutterOptions();
-
-  var called = false;
+  final binding = MockSentryNativeBinding();
 
   LoadContextsIntegration getSut(
       {Map<String, dynamic> contexts = const {
@@ -465,12 +460,7 @@ class Fixture {
           }
         ]
       }}) {
-    // ignore: deprecated_member_use
-    channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      called = true;
-      return contexts;
-    });
-
-    return LoadContextsIntegration(channel);
+    when(binding.loadContexts()).thenAnswer((_) async => contexts);
+    return LoadContextsIntegration(binding);
   }
 }