+++ flutter/test/screenshot/sentry_screenshot_widget_test.dart
@@ -1,4 +1,5 @@
 @TestOn('vm')
+library flutter_test;
 // ignore_for_file: invalid_use_of_internal_member
 
 import 'package:flutter/material.dart';